Add verbose flag
