+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "audit-urls",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "audit url statuses, including https availability",
   "main": "dist/index.js",
   "bin": {
-    "check-urls": "./dist/cli.js"
+    "audit-urls": "./dist/cli.js"
   },
   "engines": {
     "node": ">=10"
@@ -20,6 +20,7 @@
   "author": "John Teske",
   "license": "MIT",
   "dependencies": {
+    "commander": "^5.1.0",
     "got": "^11.1.3"
   },
   "devDependencies": {
