+++ src/filter.ts
@@ -1,0 +1,6 @@
+import { LinkStatus } from "./index";
+
+export const none = (v) => v;
+
+export const notOk = (v: LinkStatus) =>
+  v.error != null || v.redirect != null || v.https === "available";
